Move appropriate availability helper methods to solr document and new module
