+++ app/helpers/request_helper.rb
@@ -58,15 +58,16 @@ module RequestHelper
       query: openurl_field_values.to_query).to_s
   end
 
+  # TBD - Refactor lookup of SCRC code 
   def aeon_request_allowed(document)
     document_items = document.fetch("items_json_display", [])
-    libraries = document_items.collect { |item| library(item) }
+    libraries = document_items.collect { |item| item["permanent_library"] }
     libraries.include?("SCRC")
   end
 
   def aeon_request_button(document)
     document_items = document.fetch("items_json_display", [])
-    libraries = document_items.collect { |item| library(item) }
+    libraries = document_items.collect { |item| item["permanent_library"] }
 
     if libraries.include?("SCRC")
       button_to(t("requests.aeon_button_text"), aeon_request_url(document), class: "btn btn-primary")
