+++ app/models/solr_document.rb
@@ -2,17 +2,16 @@
 
 class SolrDocument
   include Blacklight::Solr::Document::RisFields
-  include AvailabilityHelper
   include Citable
   include JsonLogger
   include Diggable
   include QueryListable
+  include Lookupable
   include Blacklight::Solr::Document
 
   attr_accessor :logger
 
   def initialize(doc, req = nil)
-    doc[:materials_data] = materials_data
     @libkey_journals_url_thread = libkey_journals_url_thread(doc)
 
     super
@@ -84,45 +83,6 @@ class SolrDocument
     CN: "call_number_display"
   )
 
-  def materials_data
-    Proc.new {
-      @materials_data ||= alma_availability_mms_ids.map { |id|
-        # return maximum allowed item or lose items.
-        log = { type: "alma_bib_item", mms_id: id, limit: 100 }
-        do_with_json_logger(log) { Alma::BibItem.find(id, limit: 100).filter_missing_and_lost }
-      }.first
-    }
-  end
-
-  # Loads a list of books or other physical materials assoc. to record.
-  # (As apposed to an online only item)
-  def materials
-    @materials ||= materials_data[]
-      .map { |material|
-      { title: material["bib_data"]["title"],
-        barcode: barcode(material),
-        call_number: material["holding_data"]["call_number"],
-        library: library_name_from_short_code(material.library),
-        location: materials_location(material),
-        raw_library: material["item_data"]["library"]["value"],
-        raw_location: material["item_data"]["location"]["value"],
-        availability: availability_status(material) }
-        .with_indifferent_access }
-        .uniq { |material| material.except(:barcode) }
-  end
-
-  def material_from_barcode(barcode = nil)
-    materials.select { |material| material[:barcode] == barcode }.first
-  end
-
-  def barcodes
-    materials.map { |material| material[:barcode] }
-  end
-
-  def valid_barcode?(barcode = nil)
-    barcodes.include? barcode
-  end
-
   def purchase_order?
     !!self["purchase_order"]
   end
@@ -138,6 +98,50 @@ class SolrDocument
     end
   end
 
+  def document_items
+    document_items = fetch("items_json_display", [])
+    document_items.collect { |item| item }
+      .reject(&:blank?)
+      .reject { |item| missing_or_lost?(item) }
+      .reject { |item| unwanted_library_locations(item) }
+      .each { |item| item.merge! ({ library: library_name_from_short_code(library(item)) }) }
+      .each { |item| item.merge! ({ location: location_name_from_short_codes(location(item), library(item)) }) }
+      .each { |item| item.merge! ({ call_number_display: alternative_call_number(item) }) }
+  end
+
+  def document_items_grouped
+    grouped_items = document_items
+      .group_by { |item| item["library"] }
+      .transform_values { |library| library.group_by { |item| item["location"] }.sort.to_h }
+    sorted_items = grouped_items
+      .sort_by { |library, locations| [library == "Charles Library" ? 0 : 1, library] }.to_h
+      .each do |library, locations|
+        unless locations.empty?
+          locations.each do |location, items|
+            unless items.empty?
+              items.sort_by! { |item| [item["call_number_display"], item.fetch("description", "")] }
+            end
+          end
+        end
+      end
+  end
+
+  def library(item)
+    item["current_library"] ? item["current_library"] : item["permanent_library"]
+  end
+
+  def location(item)
+    item["current_location"] ? item["current_location"] : item["permanent_location"]
+  end
+
+  def call_number(item)
+    item["temp_call_number"] ? item["temp_call_number"] : item["call_number"]
+  end
+
+  def alternative_call_number(item)
+    item["alt_call_number"] ? item["alt_call_number"] : call_number(item)
+  end
+
   def libkey_journals_url
     @libkey_journals_url_thread.value&.fetch("browzineWebLink", nil)
   end
@@ -147,22 +151,6 @@ class SolrDocument
   end
 
   private
-    def barcode(item)
-      item["item_data"]["pid"]
-    end
-
-    def availability_status(item)
-      if item.in_place? && item.non_circulating?
-        "Library Use Only"
-      elsif item.in_place?
-        "Available"
-      elsif item.has_process_type?
-        Rails.configuration.process_types[item.process_type] ||
-          "Checked out or currently unavailable"
-      else
-        "Checked out or currently unavailable"
-      end
-    end
 
     def libkey_journals_url_thread(doc)
       issn = doc.fetch("issn_display", []).map { |x| x.delete("-") }.uniq.join(",")
@@ -180,5 +168,16 @@ class SolrDocument
 
     def logger
       @logger ||= Blacklight.logger
+    end
+
+    def missing_or_lost?(item)
+      process_type = item.fetch("process_type", "")
+      !!process_type.match(/MISSING|LOST_LOAN|LOST_LOAN_AND_PAID/)
+    end
+
+    def unwanted_library_locations(item)
+      library = library(item) || ""
+      location = location(item) || ""
+      !!location.match(/techserv|UNASSIGNED|intref/) || library == "EMPTY"
     end
 end
