+++ app/views/almaws/item.html.erb
@@ -11,17 +11,17 @@
 
 <div id="request-url-data-<%= @mms_id %>" class="hidden" data-requests-url="<%= request_options_path(@mms_id, @pickup_locations, @request_level) %>"></div>
 <% unless @document_availability.all?(&:empty?) %>
-<% sort_order_for_holdings(@document_availability).each do |key, locations| %>
+<% @document_availability.each do |library, locations| %>
   <% locations.each do |location, items| %>
     <div class="m-0" data-controller="availability">
       <div class="table-heading">
-        <h3 class="holding-name"><span class="library-name"><%= library_name_from_short_code(key) %></span><span class="location-name"> - <%= location_name_from_short_codes(location, key) %></span></h3>
+        <h3 class="holding-name"><span class="library-name"><%= library %></span><span class="location-name"> - <%= location %></span></h3>
         <% if summary_list(items).present? %>
           <div class="holding-summary"><p><%= summary_list(items) %></p></div>
         <% end %>
       </div>
 
-      <%= scrc_instructions(key, @document) %>
+      <%= scrc_instructions(library, @document) %>
 
       <table class="table availability-table px-2">
         <thead>
@@ -35,8 +35,8 @@
       <tbody data-long-list>
       <% items.each do |item| %>
         <tr>
-          <td><%= alternative_call_number(item) %></td>
-          <td><%= description(item) %><%= public_note(item) %></td>
+          <td><%= item.fetch("call_number_display", "") %></td>
+          <td><%= item.fetch("description", "") %><%= public_note(item) %></td>
           <td><%= material_type(item) %></td>
           <td>
             <div class="avail_status"><%= item.fetch("availability", "") %></div>
