+++ app/views/catalog/_scrc_instructions.html.erb
@@ -1,5 +1,3 @@
-<% if key == "SCRC" %>
-  <div id="scrc-instructions" class="alert alert-info pl-3" role="alert">
-    <p class="mb-0"><%= t("requests.scrc_instructions_html", href: link_to(t("requests.scrc_instructions_href"), aeon_request_url(document))) %></p>
-  </div>
-<% end %>
+<div id="scrc-instructions" class="alert alert-info pl-3" role="alert">
+  <p class="mb-0"><%= t("requests.scrc_instructions_html", href: link_to(t("requests.scrc_instructions_href"), aeon_request_url(document))) %></p>
+</div>
