+++ spec/helpers/availability_helper_spec.rb
@@ -241,24 +241,6 @@ RSpec.describe AvailabilityHelper, type: :helper do
     end
   end
 
-  describe "#description(item)" do
-    context "item includes description" do
-      let(:item) { { "description" => "v. 1" } }
-
-      it "displays description" do
-        expect(description(item)).to eq "v. 1"
-      end
-    end
-
-    context "item does NOT include description" do
-      let(:item) { {} }
-
-      it "displays nothing" do
-        expect(description(item)).to eq ""
-      end
-    end
-  end
-
   describe "#material_type(item)" do
     context "item includes physical_material_type" do
       let(:item) { { "material_type" => "RECORD" } }
@@ -299,534 +281,6 @@ RSpec.describe AvailabilityHelper, type: :helper do
 
       it "displays nothing" do
         expect(public_note(item)).to eq ""
-      end
-    end
-  end
-
-  describe "#missing_or_lost(item)" do
-    context "an item is missing" do
-      let(:item) { { "item_pid" => "23237957740003811",
-      "item_policy" => "5",
-      "permanent_library" => "AMBLER",
-      "permanent_location" => "media",
-      "current_library" => "AMBLER",
-      "current_location" => "media",
-      "call_number" => "DVD 13 A165",
-      "holding_id" => "22237957750003811",
-      "process_type" => "MISSING" }
-    }
-
-      it "correctly identifies missing items" do
-        expect(missing_or_lost?(item)).to be true
-      end
-    end
-
-    context "an item is lost" do
-      let(:item) { { "item_pid" => "23237957740003811",
-      "item_policy" => "5",
-      "permanent_library" => "AMBLER",
-      "permanent_location" => "media",
-      "current_library" => "AMBLER",
-      "current_location" => "media",
-      "call_number" => "DVD 13 A165",
-      "holding_id" => "22237957750003811",
-      "process_type" => "LOST_LOAN_AND_PAID" }
-    }
-
-      it "correctly identifies missing items" do
-        expect(missing_or_lost?(item)).to be true
-      end
-    end
-
-    context "an item is not missing or lost" do
-      let(:item) { { "item_pid" => "23237957740003811",
-      "item_policy" => "5",
-      "permanent_library" => "AMBLER",
-      "permanent_location" => "media",
-      "current_library" => "AMBLER",
-      "current_location" => "media",
-      "call_number" => "DVD 13 A165",
-      "holding_id" => "22237957750003811" }
-    }
-
-      it "correctly identifies non-missing items" do
-        expect(missing_or_lost?(item)).to be false
-      end
-    end
-  end
-
-  describe "#unwanted(item)" do
-    context "an item is missing" do
-      let(:item) { { "item_pid" => "23237957740003811",
-      "item_policy" => "5",
-      "permanent_library" => "AMBLER",
-      "permanent_location" => "techserv",
-      "current_library" => "AMBLER",
-      "current_location" => "techserv",
-      "call_number" => "DVD 13 A165",
-      "holding_id" => "22237957750003811",
-      "process_type" => "MISSING" }
-    }
-
-      it "correctly identifies techserv items" do
-        expect(unwanted_library_locations(item)).to be true
-      end
-    end
-
-    context "an item is an unwanted library" do
-      let(:item) { { "item_pid" => "23237957740003811",
-      "item_policy" => "5",
-      "permanent_library" => "AMBLER",
-      "permanent_location" => "stacks",
-      "current_library" => "EMPTY",
-      "call_number" => "DVD 13 A165",
-      "holding_id" => "22237957750003811" }
-    }
-
-      it "correctly identifies unwanted library" do
-        expect(unwanted_library_locations(item)).to be true
-      end
-    end
-
-    context "an item is not in an unwanted location" do
-      let(:item) { { "item_pid" => "23237957740003811",
-      "item_policy" => "5",
-      "permanent_library" => "AMBLER",
-      "permanent_location" => "media",
-      "current_library" => "AMBLER",
-      "current_location" => "media",
-      "call_number" => "DVD 13 A165",
-      "holding_id" => "22237957750003811" }
-    }
-
-      it "correctly identifies other locations" do
-        expect(unwanted_library_locations(item)).to be false
-      end
-    end
-  end
-
-  describe "#library(item)" do
-    context "item is in temporary library" do
-      let(:item) { { "current_library" => "RES_SHARE" } }
-
-      it "displays temporary library" do
-        expect(library(item)).to eq "RES_SHARE"
-      end
-    end
-
-    context "item is NOT in temporary library" do
-      let(:item) { { "permanent_library" => "MAIN" } }
-
-      it "displays library" do
-        expect(library(item)).to eq "MAIN"
-      end
-    end
-  end
-
-  describe "#location_name_from_short_codes(location_code, library_code)" do
-    context "location codes are converted to names using translation map" do
-      let(:library_code) { "SCRC" }
-      let(:location_code) { "rarestacks" }
-
-      it "displays location name" do
-        expect(location_name_from_short_codes(location_code, library_code)).to eq "Reading Room"
-      end
-    end
-  end
-
-  describe "#library_name_from_short_code(short_code)" do
-    context "library codes are converted to names using translation map" do
-      let(:short_code) { "MAIN" }
-      it "displays library name" do
-        expect(library_name_from_short_code(short_code)).to eq "Charles Library"
-      end
-    end
-  end
-
-  describe "#alternative_call_number(item)" do
-    context "item has an alternate call number" do
-      let(:item) { { "alt_call_number" => "alternate call number" } }
-
-      it "displays alternate call number" do
-        expect(alternative_call_number(item)).to eq "alternate call number"
-      end
-    end
-
-    context "item does NOT have an alternate call number" do
-      let(:item) { { "call_number" => "regular call number" } }
-
-      it "does NOT display alternate call number" do
-        expect(alternative_call_number(item)).to eq "regular call number"
-      end
-    end
-  end
-
-  describe "#document_availability_info(document)" do
-    context "filters out empty items" do
-      let(:document) { {} }
-
-      it "returns an empty hash" do
-        expect(document_availability_info(document)).to eq({})
-      end
-    end
-
-    context "groups by library" do
-      let(:document) { { "items_json_display" =>
-        [{ "item_pid" => "12345",
-        "item_policy" => "5",
-        "permanent_library" => "AMBLER",
-        "permanent_location" => "media",
-        "current_library" => "AMBLER",
-        "current_location" => "media",
-        "call_number" => "DVD 13 A165",
-        "holding_id" => "22237957750003811" }]
-          }
-        }
-
-      it "uses the library as a key" do
-        expect(document_availability_info(document)).to eq("AMBLER" =>
-                                                              { "media" =>
-                                                              [{ "call_number" => "DVD 13 A165",
-                                                                "current_library" => "AMBLER",
-                                                                "current_location" => "media",
-                                                                "holding_id" => "22237957750003811",
-                                                                "item_pid" => "12345",
-                                                                "item_policy" => "5",
-                                                                "permanent_library" => "AMBLER",
-                                                                "permanent_location" => "media" }] })
-      end
-    end
-
-    context "groups by location" do
-      let(:document) { { "items_json_display" =>
-        [{ "item_pid" => "23245366030003811",
-          "item_policy" => "0",
-          "permanent_library" => "KARDON",
-          "permanent_location" => "p_remote",
-          "current_library" => "KARDON",
-          "current_location" => "p_remote",
-          "call_number_type" => "0",
-          "call_number" => "N5220.K7",
-          "holding_id" => "22245366050003811",
-          "material_type" => "BOOK" },
-         { "item_pid" => "23245366040003811",
-          "item_policy" => "0",
-          "permanent_library" => "KARDON",
-          "permanent_location" => "p_remote",
-          "current_library" => "KARDON",
-          "current_location" => "p_remote",
-          "call_number_type" => "0",
-          "call_number" => "N5220.K7",
-          "holding_id" => "22245366050003811",
-          "material_type" => "BOOK" }] }
-        }
-
-      it "uses the location as a key" do
-        expect(document_availability_info(document)).to eq("KARDON" =>
-          { "p_remote" =>
-            [{ "call_number" => "N5220.K7",
-              "call_number_type" => "0",
-              "current_library" => "KARDON",
-              "current_location" => "p_remote",
-              "holding_id" => "22245366050003811",
-              "item_pid" => "23245366030003811",
-              "item_policy" => "0",
-              "material_type" => "BOOK",
-              "permanent_library" => "KARDON",
-              "permanent_location" => "p_remote" },
-              { "call_number" => "N5220.K7",
-              "call_number_type" => "0",
-              "current_library" => "KARDON",
-              "current_location" => "p_remote",
-              "holding_id" => "22245366050003811",
-              "item_pid" => "23245366040003811",
-              "item_policy" => "0",
-              "material_type" => "BOOK",
-              "permanent_library" => "KARDON",
-              "permanent_location" => "p_remote" }] })
-      end
-    end
-
-    context "groups by multiple locations" do
-      let(:document) { { "items_json_display" =>
-        [{ "item_pid" => "23280623450003811",
-        "item_policy" => "0",
-        "description" => "2014",
-        "permanent_library" => "GINSBURG",
-        "permanent_location" => "stacks",
-        "current_library" => "GINSBURG",
-        "current_location" => "stacks",
-        "call_number_type" => "2",
-        "call_number" => "WB 39 W319",
-        "holding_id" => "22280623590003811",
-        "material_type" => "ISSUE" },
-        { "item_pid" => "23280623530003811",
-        "item_policy" => "0",
-        "description" => "2001",
-        "permanent_library" => "GINSBURG",
-        "permanent_location" => "stacks",
-        "current_library" => "GINSBURG",
-        "current_location" => "stacks",
-        "call_number_type" => "2",
-        "call_number" => "WB 39 W319",
-        "holding_id" => "22280623590003811",
-        "material_type" => "ISSUE" },
-        { "item_pid" => "23468827100003811",
-        "item_policy" => "16",
-        "description" => "2020",
-        "permanent_library" => "GINSBURG",
-        "permanent_location" => "reserve",
-        "current_library" => "GINSBURG",
-        "current_location" => "reserve",
-        "call_number_type" => "2",
-        "call_number" => "WB 39 W319",
-        "holding_id" => "22280623440003811",
-        "material_type" => "BOOK" }] }
-      }
-
-      it "uses the location as a key" do
-        expect(document_availability_info(document)).to eq("GINSBURG" =>
-          { "reserve" =>
-            [{ "call_number" => "WB 39 W319",
-              "call_number_type" => "2",
-              "current_library" => "GINSBURG",
-              "current_location" => "reserve",
-              "description" => "2020",
-              "holding_id" => "22280623440003811",
-              "item_pid" => "23468827100003811",
-              "item_policy" => "16",
-              "material_type" => "BOOK",
-              "permanent_library" => "GINSBURG",
-              "permanent_location" => "reserve" }],
-            "stacks" =>
-              [{ "call_number" => "WB 39 W319",
-              "call_number_type" => "2",
-              "current_library" => "GINSBURG",
-              "current_location" => "stacks",
-              "description" => "2014",
-              "holding_id" => "22280623590003811",
-              "item_pid" => "23280623450003811",
-              "item_policy" => "0",
-              "material_type" => "ISSUE",
-              "permanent_library" => "GINSBURG",
-              "permanent_location" => "stacks" },
-              { "call_number" => "WB 39 W319",
-              "call_number_type" => "2",
-              "current_library" => "GINSBURG",
-              "current_location" => "stacks",
-              "description" => "2001",
-              "holding_id" => "22280623590003811",
-              "item_pid" => "23280623530003811",
-              "item_policy" => "0",
-              "material_type" => "ISSUE",
-              "permanent_library" => "GINSBURG",
-              "permanent_location" => "stacks" }] })
-      end
-
-      it "sorts location alphabetically by default" do
-        expect(document_availability_info(document)["GINSBURG"].keys).to eq(["reserve", "stacks"])
-      end
-
-    end
-
-
-    context "does not include missing or lost items" do
-      let(:document) { { "items_json_display" =>
-        [{ "item_pid" => "12345",
-        "item_policy" => "5",
-        "permanent_library" => "AMBLER",
-        "permanent_location" => "media",
-        "current_library" => "AMBLER",
-        "current_location" => "media",
-        "call_number" => "DVD 13 A165",
-        "process_type" => "MISSING",
-        "holding_id" => "22237957750003811" }]
-          }
-        }
-
-      it "filters out missing or lost items" do
-        expect(document_availability_info(document)).to eq({})
-      end
-    end
-  end
-
-  describe "#sort_order_for_holdings(grouped_items)" do
-    context "items are sorted by library name with Charles first" do
-      let(:grouped_items) do { "AMBLER" =>
-        { "stacks" =>
-        [{ "item_pid" => "23239405700003811",
-          "item_policy" => "0",
-          "permanent_library" => "AMBLER",
-          "permanent_location" => "stacks",
-          "current_library" => "AMBLER",
-          "current_location" => "stacks",
-          "call_number_type" => "0",
-          "call_number" => "F159.P7 C66 2003",
-          "holding_id" => "22239405730003811",
-          "availability" => "<span class=\"check\"></span>Available" }]
-        },
-        "ASRS" =>
-        { "bookbot" =>
-            [{ "item_pid" => "23239405700003811",
-              "item_policy" => "0",
-              "permanent_library" => "ASRS",
-              "permanent_location" => "bookbot",
-              "current_library" => "ASRS",
-              "current_location" => "bookbot",
-              "call_number_type" => "0",
-              "call_number" => "F159.P7 C66 2003",
-              "holding_id" => "22239405730003811",
-              "availability" => "<span class=\"check\"></span>Available" }]
-        },
-       "MAIN" =>
-       { "stacks" =>
-        [{ "item_pid" => "23239405740003811",
-          "item_policy" => "0",
-          "permanent_library" => "MAIN",
-          "permanent_location" => "stacks",
-          "current_library" => "MAIN",
-          "current_location" => "stacks",
-          "call_number_type" => "0",
-          "call_number" => "F159.P7 C66 2003",
-          "holding_id" => "22239405750003811",
-          "availability" => "<span class=\"check\"></span>Available" }]
-        }
-        }
-      end
-
-      it "returns Charles first, then Ambler" do
-        expect(sort_order_for_holdings(grouped_items).keys).to eq(["MAIN", "ASRS", "AMBLER"])
-      end
-
-      it "returns ASRS second" do
-        expect(sort_order_for_holdings(grouped_items).keys).to eq(["MAIN", "ASRS", "AMBLER"])
-      end
-    end
-
-    context "items in Kardon sort by Remote Storage, not KARDON" do
-      let(:grouped_items) do {
-      "KARDON" =>
-      { "p_remote" =>
-        [{ "item_pid" => "23243718620003811",
-         "item_policy" => "0",
-         "permanent_library" => "KARDON",
-         "permanent_location" => "p_remote",
-         "current_library" => "KARDON",
-         "current_location" => "p_remote",
-         "call_number_type" => "0",
-         "call_number" => "N6853.S49 A4 2001",
-         "holding_id" => "22243718630003811",
-         "availability" => "<span class=\"check\"></span>Available" }]
-      },
-      "MAIN" =>
-      { "stacks" =>
-         [{ "item_pid" => "23243718640003811",
-         "item_policy" => "0",
-         "permanent_library" => "MAIN",
-         "permanent_location" => "stacks",
-         "current_library" => "MAIN",
-         "current_location" => "stacks",
-         "call_number_type" => "0",
-         "call_number" => "N6853.S49 A4 2001",
-         "holding_id" => "22243718650003811",
-         "availability" => "<span class=\"check\"></span>Available" }]
-      }
-      }
-      end
-
-      it "returns Main before Kardon" do
-        expect(sort_order_for_holdings(grouped_items).keys).to eq(["MAIN", "KARDON"])
-      end
-    end
-
-    context "Items are ordered by call number after location" do
-      let(:grouped_items) do
-        { "MAIN" =>
-          { "stacks" =>
-          [{ "item_pid" => "23242235660003811",
-          "item_policy" => "12",
-          "description" => "1992-94",
-          "permanent_library" => "MAIN",
-          "permanent_location" => "stacks",
-          "current_library" => "MAIN",
-          "current_location" => "stacks",
-          "call_number_type" => "0",
-          "call_number" => "MT655.P45x",
-          "holding_id" => "22242235730003811",
-          "availability" => "<span class=\"check\"></span>Library Use Only" },
-         { "item_pid" => "23242235720003811",
-          "item_policy" => "12",
-          "description" => "1983-1986",
-          "permanent_library" => "MAIN",
-          "permanent_location" => "stacks",
-          "current_library" => "MAIN",
-          "current_location" => "stacks",
-          "call_number_type" => "0",
-          "call_number" => "HF5006 .I614",
-          "holding_id" => "22242235730003811",
-          "availability" => "<span class=\"check\"></span>Library Use Only" },
-         { "item_pid" => "23242235710003811",
-          "item_policy" => "12",
-          "description" => "1987-89",
-          "permanent_library" => "MAIN",
-          "permanent_location" => "stacks",
-          "current_library" => "MAIN",
-          "current_location" => "stacks",
-          "call_number_type" => "0",
-          "call_number" => "AC1 .G72",
-          "holding_id" => "22242235730003811",
-          "availability" => "<span class=\"check\"></span>Library Use Only" }]
-        }
-        }
-      end
-
-      it "returns copies for each library by call number" do
-        sorted_call_numbers = sort_order_for_holdings(grouped_items)["MAIN"]["stacks"].map { |item| alternative_call_number(item) }
-        expect(sorted_call_numbers).to eq(["AC1 .G72", "HF5006 .I614", "MT655.P45x"])
-      end
-    end
-
-    context "Items are ordered by description after call number" do
-      let(:grouped_items) do
-        { "MAIN" =>
-          { "stacks" =>
-          [{ "item_pid" => "23242235660003811",
-          "item_policy" => "12",
-          "description" => "v.55, no.5 (Nov. 2017)",
-          "permanent_library" => "MAIN",
-          "permanent_location" => "stacks",
-          "current_library" => "MAIN",
-          "current_location" => "stacks",
-          "call_number_type" => "0",
-          "call_number" => "MT655.P45x",
-          "holding_id" => "22242235730003811" },
-         { "item_pid" => "23242235720003811",
-          "item_policy" => "12",
-          "description" => "v.53 (2016)",
-          "permanent_library" => "MAIN",
-          "permanent_location" => "stacks",
-          "current_library" => "MAIN",
-          "current_location" => "stacks",
-          "call_number_type" => "0",
-          "call_number" => "MT655.P45x",
-          "holding_id" => "22242235730003811" },
-         { "item_pid" => "23242235710003811",
-          "item_policy" => "12",
-          "description" => "v.42 (2004)",
-          "permanent_library" => "MAIN",
-          "permanent_location" => "stacks",
-          "current_library" => "MAIN",
-          "current_location" => "stacks",
-          "call_number_type" => "0",
-          "call_number" => "MT655.P45x",
-          "holding_id" => "22242235730003811" }]
-        }
-        }
-      end
-
-      it "returns copies for each library by description" do
-        sorted_descriptions = sort_order_for_holdings(grouped_items)["MAIN"]["stacks"].map { |item| description(item) }
-        expect(sorted_descriptions).to eq(["v.42 (2004)", "v.53 (2016)", "v.55, no.5 (Nov. 2017)"])
       end
     end
   end
